requirements.txt
#git+https://github.com/kshubham506/Telethon@b1984615afc684e1182ff82674dcfc0a43ed5cd9
telethon==1.27.0
asyncio==3.4.3
requests==2.25.1
python-binance==1.0.16
colorama==0.4.4
requests-toolbelt==0.10.1
pyrebase4
numpy==1.24.1
python-dotenv==0.20.0
discord.py-self==1.9.2
discord.py==2.3.2
nest-asyncio==1.5.6
<<<<<<< HEAD
dotmap
munch
jsonpickle
pandas
=======
dotmap==1.3.30
jsonpickle==3.0.1
munch==2.5.0
schedule==1.2.1
pandas==2.0.1
>>>>>>> new-main
